move migration functions to own file, exporting to keep migratePrefs where it was
